refactor(agents): split monolithic tools.ts into per-domain files
MON-557 — break 786 LOC tools.ts into setup/services/prices/meters/benefits/coupons.
Registry abstraction preserved (justified by MON-558 support skill incoming).
No file >300 LOC.
